Move analyseOne to Analysis module
